+++ Notemania/src/Beatmap.cpp
@@ -1,0 +1,6 @@
+#include "Beatmap.hpp"
+
+namespace noma
+{
+    
+} // namespace noma
